pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "API service for task mgmt"
authors = [
    {name = "Ayikoru Alima",email = "alima.ayikoru@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
